added null check for expiration date

+++ EatSenseRestWS/src/main/java/net/eatsense/auth/AccessTokenFilter.java
@@ -79,7 +79,7 @@ public class AccessTokenFilter implements ContainerRequestFilter {
 				throw new IllegalAccessException("Access token invalid, please re-authenticate.");
 			}
 			
-			if(accessToken.getExpires().before(new Date()) ) {
+			if( accessToken.getExpires() != null &&  accessToken.getExpires().before(new Date()) ) {
 				accessTokenRepo.delete(accessToken);
 				logger.info("Deleted expired access token for account {}", accessToken.getAccount());
 				throw new IllegalAccessException("Access token exired, please re-authenticate.");
